% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle.hough.r
\name{circle.hough}
\alias{circle.hough}
\title{Estimate parameters of a circle using Hough Circle Transform}
\usage{
circle.hough(
  im,
  fw = 2,
  qt = 0.995,
  excl = 5,
  rmin = min(dim(im))/4,
  rmax = min(dim(im))/2,
  rstep = 1,
  dtheta_max = 0.5,
  dtheta_step = 0.05,
  nn = 7,
  plots = TRUE,
  details = FALSE
)
}
\arguments{
\item{im}{The image to find a circle in (a modulation estimate is best)}

\item{fw}{Size of Gaussian blur to smooth image}

\item{qt}{Threshold to accept strong edge candidate}

\item{excl}{Number of pixels to exclude around edge of frame as candidates}

\item{rmin}{Minimum circle radius}

\item{rmax}{Maximum circle radius}

\item{rstep}{step size in constructing lookup table}

\item{dtheta_max}{maximum assumed error in gradient direction}

\item{dtheta_step}{increment for dtheta}

\item{plots}{plot?}

\item{details}{Return extra details?}

\item{number}{of nearest neighbors for alternate calculation}
}
\value{
If details is FALSE a named list with the circle parameters
}
\description{
Uses a portion of the canny algorithm to find candidate edge points
and the direction of the gradient at those points, then uses
Hough Circle Transform to estimate circle parameters.
}
\details{
The Hough transform section first creates a lookup table of candidate
radii and center points, then for each candidate edge point calculates
potential centers along a fan of rays near the gradient direction.
An inner join then finds matches in the lookup table and increments
an accumulator vector. Highest vote at the end wins.
}
\note{
This is experimental and can be very slow. A good guess for the radius is very helpful.
Experimental feature: find the nn nearest neighbors of the selected trio of parameters
and calculate a vote weighted mean. This is returned as \code{rxy_alt} if details is TRUE.
}
\examples{
example("psifit", package="zernike", ask=FALSE)
X11()
cp2 <- circle.hough(tfit$mod, rmin=round(tfit$cp$rx)-10, rmax=round(tfit$cp$rx)+10)
}
\seealso{
[circle.pars()], [pupil.pars()]
}
