% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{conv2_sep_complex}
\alias{conv2_sep_complex}
\title{2D separable convolution - complex matrix}
\usage{
conv2_sep_complex(X, kernel)
}
\arguments{
\item{X}{a complex valued matrix.}

\item{kernel}{}
}
\value{
a complex valued  matrix with the same dimension as X.
}
\description{
2D convolution with a separable kernel
}
\details{
Having a separable kernel implies that a
direct 2D convolution can be executed as a series
of 1D column- and row-wise convolutions.
This \emph{may} be faster than a direct 2D convlution or
using FFT's, and certainly benefits from multithreading.
The input matrix is implicitly zero padded, which differs
from an fft based convolution, where the input is implicitly
extended to infinity in all directions, therefore the outputs
will in general differ
}
\seealso{
\code{\link[=conv2_sep_real]{conv2_sep_real()}}
}
