% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convolve2d}
\alias{convolve2d}
\title{2D convolution - real matrix}
\usage{
convolve2d(X, kernel)
}
\arguments{
\item{X}{a real valued matrix.}

\item{kernel}{a real valued matrix}
}
\value{
a real matrix with the same dimension as X.
}
\description{
General 2D dircet convolution
}
\details{
General 2D direct convolution with an arbitrary matrix kernel.
This is a wrapper for the Armadillo function \code{conv2}.
}
