% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{d_of_g}
\alias{d_of_g}
\title{Derivative of Gaussian filter}
\usage{
d_of_g(X, sigma)
}
\arguments{
\item{X}{a matrix}

\item{sigma}{standard deviation in pixels of the Gaussian kernel}
}
\value{
A named list with components Dx, Dy, Dxy, and the calculated kernel

Calls arma::conv2(). Called by \code{\link[=circle.auto]{circle.auto()}}.
}
\description{
2D filter that combines smoothing and (central) differencing.
}
