% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fft_cx}
\alias{fft_cx}
\title{2D FFT - complex valued input}
\usage{
fft_cx(X)
}
\arguments{
\item{X}{a complex valued matrix}
}
\value{
A complex valued matrix with the discrete Fourier transformed input.
}
\description{
2D FFT wrapping Armadillo's \code{fft2}
}
\seealso{
\code{\link[=fft]{fft()}}
}
