% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fft_pad}
\alias{fft_pad}
\title{Wrappers for Armadillo 2D fft's with padding}
\usage{
fft_pad(X, npad)
}
\description{
2D ffts with real and complex inputs
}
\details{
@param X a real or complex valued metrix
@param npad size to pad to (should be highly composite)

@returns A complex valued transformed matrix of size npad x npad.

@details For convenience there are versions for
real or complex inputs. Armadillo's \code{fft2} function
is poorly suited to matrixes of sizes other than
highly composite. These functions zero pad the input
and compute the transform of the padded matrix.
\code{npad} obviously should be highly composite or a power of two.
}
