% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fftshift}
\alias{fftshift}
\title{2D fftshift}
\usage{
fftshift(X)
}
\arguments{
\item{X}{a real or complex valued matrix.}
}
\value{
a complex matrix with the same dimension as X.
}
\description{
Shift quadrants of a matrix
}
\details{
This swaps the quadrants of
a matrix to put the (0,0) element
in the center. The most common use is
to center the DC component of a Fourier Transformed matrix
for visualization or to simplify other matrix operations
}
\examples{
X <- matrix(1:16, 4, 4)
XS <- fftshift(X)
XS
ifftshift(XS)
}
