% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gblur_complex}
\alias{gblur_complex}
\title{2D Gaussian blur of a complex valued matrix}
\usage{
gblur_complex(X, sigma)
}
\arguments{
\item{X}{a complex valued matrix}

\item{sigma}{standard deviation in pixels of the Gaussian kernel}
}
\value{
A complex valued matrix with the same dimensions as the input

Performs direct convolution using \code{\link[=conv2_sep_complex]{conv2_sep_complex()}} with a Gaussian kernel
sized to span 5 standard deviations. Preliminary benchmarking indicates this
is no faster than performing separate convolutions of the real and imaginary
components and combining the results.
}
\description{
2D Gaussian blur of a complex matrix
}
