% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gblur_fft}
\alias{gblur_fft}
\title{2D Gaussian blur}
\usage{
gblur_fft(X, sigma)
}
\arguments{
\item{X}{a real valued matrix}

\item{sigma}{standard deviation in pixels of the Gaussian kernel}
}
\value{
the smoothed input matrix

Numerical experiments suggest that it's almost always faster to use
the direct convolution based Gaussian blur in \code{\link[=gblur]{gblur()}}.
}
\description{
2D Gaussian blur -- FFT based version
}
