% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gol_welsch}
\alias{gol_welsch}
\title{Golub-Welsch method to find quadrature points and weights for Gauss-Legendre quadrature}
\usage{
gol_welsch(eps, qwts)
}
\arguments{
\item{eps}{obstruction fraction 0 <= eps < 1}

\item{qwts}{an input R vector with length equal to the number of quatrature points.}
}
\value{
a vector of quadrature nodes the same length as \code{qwts} in the open interval
(eps^2, 1). \code{qwts} will be overwritten with the quadrature weights.
}
\description{
Calculates the nodes and weights for Legendre polynomials shifted
to the interval (eps^2, 1).
}
\details{
If N is the maximum polynomial order to be evaluated qwts should be at least
of length N/2 + 1. Quadratures will then be nominally exact.
}
\seealso{
Called by \code{\link[=zapm]{zapm()}} and \code{\link[=zapm_iso]{zapm_iso()}}.
}
