% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ifft}
\alias{ifft}
\title{2D Inverse FFT - complex valued input}
\usage{
ifft(X)
}
\arguments{
\item{X}{a complex valued matrix}
}
\value{
A complex valued matrix with the inverse discrete Fourier transformed input.

In contrast to some implementations this returns the \emph{scaled} inverse transform.
}
\description{
2D Inverse FFT wrapping Armadillo's \code{ifft2}
}
\seealso{
\code{\link[=fft]{fft()}}
}
