% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ifft_real}
\alias{ifft_real}
\title{2D Inverse FFT - real valued output}
\usage{
ifft_real(X)
}
\arguments{
\item{X}{a complex valued matrix}
}
\value{
A real valued matrix with the real component of the inverse discrete Fourier transformed input.

In many cases where a real valued matrix is transformed, some operations performed in the Fourier domain,
and then inverse transformed the output will be real but with a small imaginary component.
This function simply discards the imaginary component. No check is made that this is
actually appropriate.
}
\description{
2D Inverse FFT wrapping Armadillo's \code{ifft2}
}
\seealso{
\code{\link[=fft]{fft()}}
}
