% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ifftshift}
\alias{ifftshift}
\title{2D ifftshift}
\usage{
ifftshift(X)
}
\arguments{
\item{X}{a real or complex valued matrix.}
}
\value{
a complex matrix with the same dimension as X.
}
\description{
Inverse fftShift
}
\details{
Inverts the fftshift operation,
swapping quadrants of the input matrix.
}
\examples{
X <- matrix(1:16, 4, 4)
XS <- fftshift(X)
XS
ifftshift(XS)
}
\seealso{
\code{\link[=fftshift]{fftshift()}}.
}
