% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zernikes.r
\name{makezlist.iso}
\alias{makezlist.iso}
\title{Construct list of ZP indexes in ISO/ANSI sequence with sine terms first}
\usage{
makezlist.iso(maxorder = 12)
}
\arguments{
\item{maxorder}{maximum radial and azimuthal order}
}
\value{
a named list with n=radial indexes, m=azimuthal, and t indicating
trig function to apply
}
\description{
Construct list of ZP indexes in ISO/ANSI sequence with sine terms first
}
\examples{
zlist.iso <- makezlist.iso(maxorder=6)
zlist <- makezlist(0, 6)
}
