% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{norm_zpm}
\alias{norm_zpm}
\title{Normalize matrix of Zernike polynomial values.}
\usage{
norm_zpm(uzpm, maxorder)
}
\arguments{
\item{uzpm}{matrix of Zernike polynomial values}

\item{maxorder}{the maximum radial order.}
}
\value{
matrix in orthonormal form.
}
\description{
Convert a matrix of Zernike polynomial values from
unit scaled to unit variance aka orthonormal form.
}
\details{
This is intended only for ISO/ANSI ordered matrices. The
only check performed is that the number of columns in the
matrix matches the expected number given by the argument
\code{maxorder}.
This is called by \code{\link[=gradzpm_arma]{gradzpm_arma()}} and \code{\link[=zpm_cart]{zpm_cart()}}
if \code{unit_variance} is set to \code{true} in the respective
function calls.
}
