% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil.r
\name{PVr}
\alias{PVr}
\title{Zygo's "robust" PV}
\usage{
PVr(wf.zfit, wf.residual)
}
\arguments{
\item{wf.zfit}{matrix containing the smoothed Zernike fit wavefront}

\item{wf.residual}{matrix of the difference between the raw wavefront
and the Zernike fit. These values are returned by \code{\link[=wf_net]{wf_net()}}}
}
\value{
the estimated PVr
}
\description{
A peak to valley error estimate that reduces the effect of noise and artifacts
}
\details{
no check is performed on the wavefronts, so it's the user's
responsibility to make sure these come from the same source
}
\references{
Evans, C. (2009) Optical Engineering 48(4), 43605.
\url{https://doi.org/10.1117/1.3119307}
}
