% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{q_uw}
\alias{q_uw}
\title{Compiled code via Rcpp for quality guided phase unwrapping}
\usage{
q_uw(nr, nc, phase, qual)
}
\arguments{
\item{nr}{number of rows in phase matrix}

\item{nc}{number of columns in phase matrix}

\item{phase}{phase matrix converted to vector}

\item{qual}{quality matrix converted to vector}
}
\value{
a vector with the unwrapped phase
}
\description{
Called by [qpuw()] for fast quality guided phase unwrapping
}
\details{
This is called by [qpuw()] but is also user callable.
  Wrapped phase values are divided by \code{2*pi} before input
  making the input values in the range [-1/2, 1/2).
}
\seealso{
[qpuw()], [idiffpuw()]
}
\author{
M.L. Peck (mlpeck54 -at- gmail.com)  
  with valuable programming advice from Steve Koehler
}
