% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zapm_direct.r
\name{rzernike_ann_direct}
\alias{rzernike_ann_direct}
\title{Radial Zernike annular polynomials from formulas}
\usage{
rzernike_ann_direct(rho, eps, j)
}
\arguments{
\item{rho}{a vector of radial coordinates.}

\item{eps}{the obstruction fraction 0 <= eps < 1.}

\item{j}{index of the value to fetch}
}
\value{
A vector of radial Zernike annular polynomial values at the coordinates \code{rho} for index \code{j}
}
\description{
The first 15 radial Zernike annular polynomials copied from
Table 1 of Mahajan (1994)
}
\details{
This function is included for testing and reference only.
}
\references{
Mahajan, V. N. (1994). Zernike Annular Polynomials and Optical Aberrations
of Systems with Annular Pupils. \emph{Supplement to Applied Optics} \strong{33}, 8125-8127.
}
\seealso{
This function is called by \code{\link[=zapm_direct]{zapm_direct()}} and \code{\link[=zapm_iso_direct]{zapm_iso_direct()}}.
Complete sequences to arbitrary order can be calculated with \code{\link[=rzernike_ann]{rzernike_ann()}}
and \code{\link[=rzernike_ann_128]{rzernike_ann_128()}}.
}
