% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate.wf.r
\name{addfit}
\alias{addfit}
\alias{separate.wf}
\title{Separate wavefronts}
\usage{
addfit(..., th = 0, zcm = NULL, theta = numeric(0))

separate.wf(
  zcm,
  theta,
  maxorder = 14,
  nrow = nrow.default,
  ncol = ncol.default,
  cp = cp.default
)
}
\arguments{
\item{...}{one or more wavefront fits as returned by \link{psifit}, \link{vortexfit}, \link{fftfit}, or \link{wf_net}.}

\item{th}{the orientation angle(s) of object under test, in degrees.}

\item{zcm}{Zernike coefficient matrix from the \code{zcoef.net} entry in the wavefront fits, minus the first 3 elements.}

\item{theta}{vector of angles from addfit.}

\item{maxorder}{the maximum Zernike polynomial order of the fits.}

\item{nrow}{number of rows in the reconstructed wavefronts.}

\item{ncol}{number of columns in the reconstructed wavefronts.}

\item{cp}{list of values describing the location and size of the wavefront as returned by \link{circle.pars} or \link{pupil.pars}.}

\item{zcb}{a data frame with zernike coefficients and standard errors of estimates for intrinsic and instrumental aberrations.}

\item{sumstats}{a data frame with summary statistics describing the fits to each set of coefficients.}

\item{wf.mirror}{the estimated derotated wavefront, stored in a matrix of size nrow x ncol with class \code{pupil}.}
}
\value{

}
\description{
Separate a set of wavefronts measured at different orientations into "polished in" and
instrumental or "test stand" components.
}
\details{
The two functions \code{addfit()} and \code{separate.wf()} work together to provide a rudimentary project management capability when multiple
optical tests have been run on an optical system rotated to one or more orientations.
All \code{addfit()} does is extract the net Zernike coefficients in the list of values returned by
\link{psifit}, \link{vortexfit}, \link{fftfit}, or \link{wf_net}. It also accepts one or more orientation angles in degrees.

The odd parameter list is to provide some flexibility in data entry. For example if several fits are available they could be entered as a group
with the \code{theta}s entered as a vector the same length as the number of fits. Alternately fits could be entered one at a time.
If the latter strategy is followed be sure to recycle the variable name that the return is assigned to. Also, the Zernike polynomial fits
should be made to the same polynomial order. This isn't checked and will surely cause an error if different fit orders are used.

\code{separate.wf()} makes use of the properties of Zernike polynomials under rotations to disentangle the contributions from the "mirror"
and the instrument or test stand to the extent possible. Least squares fits are performed for each non-axysmmetric aberration and
some possibly useful summary statistics from the fits are returned in the data frames \code{zcb} and \code{sumstats}.
}
\section{Functions}{
\itemize{
\item \code{addfit}: Add a wavefront fit to tracking list

\item \code{separate.wf}: Separate wavefronts
}}

