% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wf_net.r
\name{summary.wf_fitted}
\alias{summary.wf_fitted}
\title{Methods for class "wf_fitted"}
\usage{
\method{summary}{wf_fitted}(wffit, digits = 3)
}
\arguments{
\item{wffit}{the return values from one of the fringe analysis routines}

\item{digits}{number of digits to display in print and summary methods}

\item{...}{values passed to \code{\link[=plot.pupil]{plot.pupil()}}}
}
\value{
print method returns data frame with Zernike coefficients
}
\description{
Summary, print, and plot methods for the returned list of values
from \code{\link[=psifit]{psifit()}}, \code{\link[=fftfit]{fftfit()}}, or \code{\link[=vortexfit]{vortexfit()}}
}
