% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wf_net.r
\name{summary.wf_zfit}
\alias{summary.wf_zfit}
\title{Methods for class "wf_zfit"}
\usage{
\method{summary}{wf_zfit}(wffit, digits = 3, printnow = TRUE)
}
\arguments{
\item{wffit}{the return values from one of the fringe analysis routines or \code{\link[=wf_net]{wf_net()}}}

\item{digits}{number of digits to display in print and summary methods}

\item{printnow}{send output to console?}

\item{...}{values passed to \code{\link[=plot.pupil]{plot.pupil()}}}
}
\value{
summary and print methods return data frame with wavefront summaries and Zernike coefficients
}
\description{
Summary, print, plot, report, and invert methods for the returned list of values
from \code{\link[=psifit]{psifit()}}, \code{\link[=fftfit]{fftfit()}}, or \code{\link[=vortexfit]{vortexfit()}}
}
\details{
The report method combines the plot, summary, and print methods into an html document and displays it in a browser.
The invert method negates the values of wavefronts and Zernike coefficients and returns the adjusted input
}
