% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{zapm_cart}
\alias{zapm_cart}
\title{Zernike Annular polynomials}
\usage{
zapm_cart(x, y, maxorder = 12L)
}
\arguments{
\item{x}{a vector of x coordinates for points on a unit disk.}

\item{y}{a vector of y coordinates.}

\item{maxorder}{the maximum radial polynomial order (defaults to 12).}
}
\value{
a matrix of approximate Zernike Annular polynomial values evaluated at the input
Cartesian coordinates and all radial and azimuthal orders from
0 through \code{maxorder}.
}
\description{
Calculate approximate Zernike Annular polynomial values in
ISO/ANSI sequence for a set of Cartesian coordinates.
}
\details{
Uses QR decomposition applied separately to each azimuthal order m to orthogonalize a matrix
of Zernike polynomials. This closely approximates annular Zernikes for a large enough set of coordinates.

Note the coordinates must be uniformly spaced for this to produce the intended values.
}
