\name{convolve2d}
\alias{convolve2d}
\title{2D convolution}
\description{
  General 2D convolution
}
\usage{
convolve2d(X, kernel)
}
\arguments{
  \item{X}{A matrix representing an image}
  \item{kernel}{the convolution kernel}
}
\details{
}
\value{
  The filtered matrix X.
}
\author{M.L. Peck \email{mpeck1@ix.netcom.com}}
\seealso{
  \code{\link{gblur}}.
}
\keyword{mathematics}
