% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fft}
\alias{fft}
\title{2D FFT}
\usage{
fft(X)
}
\arguments{
\item{X}{a real valued matrix}
}
\value{
A complex valued matrix with the discrete Fourier transformed input.

This is just a wrapper for Armadillo's \code{fft2}
function. It works well for dimensions that are
\emph{highly composite}, not just powers of 2 as
implied by the online documentation. \code{X} need not
be a square matrix.
}
\description{
2D FFT wrapping Armadillo's \code{fft2}
}
\references{
Conrad Sanderson and Ryan Curtin.
Armadillo: An Efficient Framework for Numerical Linear Algebra.
International Conference on Computer and Automation Engineering, 2025.
\if{html}{\out{<br>}}
Conrad Sanderson and Ryan Curtin.
Practical Sparse Matrices in C++ with Hybrid Storage and Template-Based Expression Optimisation.
Mathematical and Computational Applications, Vol.24, No.3, 2019.
}
