% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fft_pad}
\alias{fft_pad}
\title{2D FFT with padding}
\usage{
fft_pad(X, npad)
}
\arguments{
\item{X}{a real valued matrix}

\item{npad}{size to pad to}
}
\value{
A complex valued matrix with the discrete Fourier transformed input.

This is provided for cases where the input matrix has one or both dimensions
that are \emph{not} highly composite (that is, have prime factors that are larger than 5).
\code{npad} itself should be highly composite. This isn't checked. The R function \code{\link[=nextn]{nextn()}}
can be used to find a suitable size.
}
\description{
2D FFT wrapping Armadillo's \code{fft2}
}
\seealso{
\code{\link[=fft]{fft()}}
}
