% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gblur}
\alias{gblur}
\title{2D Gaussian blur}
\usage{
gblur(X, sigma)
}
\arguments{
\item{X}{a matrix}

\item{sigma}{standard deviation in pixels of the Gaussian kernel}
}
\value{
A matrix with the same dimensions as the input

Performs direct convolution using \code{\link[=conv2_sep_real]{conv2_sep_real()}} with a Gaussian kernel
sized to span 5 standard deviations.
}
\description{
2D Gaussian blur of a real matrix
}
