% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ifft}
\alias{ifft}
\title{Wrappers for Armadillo 2D inverse fft's}
\usage{
ifft(X)
}
\description{
2D inverse ffts with complex inputs
}
\details{
For convenience there are versions for
both real and complex outputs. Note that a round trip
real FFT -> inverse FFT will produce a numerically
complex valued output with negligible imaginary component.
The function \code{\link[=ifft_real]{ifft_real()}} is intended for this case.
These functions
work well if the input dimensions are highly composite,
that is have no factors larger than 5.

@param X a complex valued metrix.

@returns A complex or real valued transformed matrix.
}
