% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rzernike_ann}
\alias{rzernike_ann}
\title{Radial Zernike Annular polynomials}
\usage{
rzernike_ann(rho, eps, n, m, xq, qwts)
}
\arguments{
\item{rho}{a vector of radial coordinates.}

\item{eps}{the obstruction fraction 0 <= eps < 1.}

\item{n}{the maximum radial order required}

\item{m}{azimuthal order}

\item{xq}{nodes for Gauss-Legendre quadrature}

\item{qwts}{weights for Gauss-Legendre quadrature}
}
\value{
A length(rho) x (n-m)/2+1 column matrix of radial Zernike Annular polynomial values evaluated at the input
radial coordinates. The radial indexes are in increasing order from m, m+2, ..., n.
}
\description{
Create a matrix of Radial Zernike annular polynomial values
for a given azimuthal order \code{m} and all radial orders up to \code{n}.
}
\details{
To the author's knowledge no recurrence relations for radial Zernike annular polynomials
have been published, even though several are well known for the closely related Zernike circle polynomials.
The m=0 polynomials representing axially symmetric aberrations however are just shifted Legendre polynomials
with an easily derived recurrence relation. For the m>0 polynomials this routine calculates
the recurrence coefficients iteratively using a method attributed to Stieltjes.
}
\references{
Gautschi, W. 1982, "On Generating Orthogonal Polynomials", SIAM J. Sci. Stat. Comput. vol. 3, no.3, 289-317.
Mahajan, V.N. 1981, "Zarnike annular polynomials...", JOSA, vol. 71, no. 1, 75-85.
Mahajan, V.N. 1994, "Zernike annular polynomials...", Suppl. Applied Optics, vol. 5, No. 11, 8125-8128.
}
\seealso{
This function is called by \code{\link[=zapm]{zapm()}} and \code{\link[=zapm_iso]{zapm_iso()}}.
}
