% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zapm_direct.r
\name{zapm_iso_direct}
\alias{zapm_iso_direct}
\title{Zernike Annular polynomials from formulas}
\usage{
zapm_iso_direct(rho, theta, eps)
}
\arguments{
\item{rho}{a vector of radial coordinates with eps <= rho <= 1.}

\item{theta}{a vector of angular coordinates, in radians.}

\item{eps}{the obstruction fraction 0 <= eps < 1.}
}
\value{
a \verb{length(rho) x 28} matrix of Zernike annular polynomial values.
}
\description{
Create a matrix of Zernike Annular polynomial values
complete through radial and azimuthal orders 6
in ISO/ANSI sequence.
}
\details{
The values are from published formulas. This function is
included for testing and reference only.
}
\seealso{
Calls \code{\link[=rzernike_ann_direct]{rzernike_ann_direct()}} for radial Zernike annular values.
Use \code{\link[=zapm_iso]{zapm_iso()}} or \code{\link[=zapm_iso_128]{zapm_iso_128()}} for complete sequences
of values to arbitrary order.
}
