% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{zpm_cart}
\alias{zpm_cart}
\title{Zernike polynomials}
\usage{
zpm_cart(x, y, maxorder = 12L, unit_variance = TRUE)
}
\arguments{
\item{x}{a vector of x coordinates for points on a unit disk.}

\item{y}{a vector of y coordinates.}

\item{maxorder}{the maximum radial polynomial order (defaults to 12).}

\item{unit_variance}{logical: return with orthonormal scaling? (default \code{false})}
}
\value{
a matrix of Zernike polynomial values evaluated at the input
Cartesian coordinates and all radial and azimuthal orders from
0 through \code{maxorder}.
}
\description{
Calculate Zernike polynomial values in
ISO/ANSI sequence for a set of Cartesian coordinates.
}
\details{
This is the same algorithm and essentially the same code as \code{\link[=gradzpm_cart]{gradzpm_cart()}}
except directional derivatives aren't calculated.
}
